// src/types/user-type.ts
export interface User {
  userId: number;
  armedForce: string;
  userRank: string;
  userName: string;
  serviceNumber: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNo: string;
  avatarUrl: string;
  lastLogin: string;
  status: number;
  created: string;
  postsCount: number;
}